update tasks and rename main workflow
